rosmongolog_mp: per topic graphs, pimped graphs
Graphs can now be created per topic, which can give an insight to what's
going on. But since this can be quite demanding it is optional and must
be enabled with the --graph-topics option. The --graph-dir option can be
used to name a directory in which to create the graphs. Pimped the graphs
a little, adding vertical description labels and using slope mode.